Dives are sent in POST body, not in URL !

+++ src/org/subsurface/GpsActivity.java
@@ -17,14 +17,19 @@
  */
 package org.subsurface;
 
-import java.net.URLEncoder;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import org.apache.http.HttpResponse;
+import org.apache.http.NameValuePair;
+import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.impl.client.DefaultHttpClient;
+import org.apache.http.message.BasicNameValuePair;
+import org.apache.http.protocol.HTTP;
 import org.subsurface.dao.DbAdapter;
 import org.subsurface.dao.DiveLocationLogDao;
 import org.subsurface.model.DiveLocationLog;
@@ -62,35 +67,27 @@ public class GpsActivity extends Activity {
 	private DiveLocationLogDao locationDao;
 
 	/**
-	 * Builds URL for a log.
+	 * Builds action for a log.
 	 * @param log log to transmit
-	 * @return URL to call, or null if settings are not available
+	 * @return action to call
 	 */
-	public String getSendUrl(DiveLocationLog log) {
+	private HttpPost getHttpPost(DiveLocationLog log) throws Exception {
 		SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
 		String destUrl = prefs.getString("destination_url", null);
 		String userId = prefs.getString("user_id", null);
-
-		String url = null;
-		if (destUrl != null && userId != null) { // Preferences already set
-			Date logDate = new Date(log.getTimestamp());
-			String date = new SimpleDateFormat("yyyy-MM-dd").format(logDate);
-			String hour = new SimpleDateFormat("HH:mm").format(logDate);
-			String name = log.getName();
-			try {
-				name = URLEncoder.encode(log.getName(), "UTF-8");
-			} catch (Exception ignored) {}
-
-			url = new StringBuilder()
-					.append(destUrl)
-					.append("/?login=").append(userId)
-					.append("&dive_latitude=").append(log.getLatitude())
-					.append("&dive_longitude=").append(log.getLongitude())
-					.append("&dive_date=").append(date)
-					.append("&dive_time=").append(hour)
-					.append("&dive_name=").append(name).toString();
-		}
-		return url;
+		HttpPost post = new HttpPost(destUrl);
+		Date logDate = new Date(log.getTimestamp());
+		String date = new SimpleDateFormat("yyyy-MM-dd").format(logDate);
+		String hour = new SimpleDateFormat("HH:mm").format(logDate);
+		List<NameValuePair> nameValuePairs = new ArrayList<NameValuePair>();
+		nameValuePairs.add(new BasicNameValuePair("login", userId));
+		nameValuePairs.add(new BasicNameValuePair("dive_latitude", Double.toString(log.getLatitude())));
+		nameValuePairs.add(new BasicNameValuePair("dive_longitude", Double.toString(log.getLongitude())));
+		nameValuePairs.add(new BasicNameValuePair("dive_date", date));
+		nameValuePairs.add(new BasicNameValuePair("dive_time", hour));
+		nameValuePairs.add(new BasicNameValuePair("dive_name", log.getName()));
+		post.setEntity(new UrlEncodedFormEntity(nameValuePairs, HTTP.UTF_8));
+		return post;
 	}
 
     @Override
@@ -168,12 +165,13 @@ public class GpsActivity extends Activity {
 							public void run() {
 								locationLog.setLocation(location);
 								locationLog.setTimestamp(System.currentTimeMillis());
-								String url = getSendUrl(locationLog);
 								try {
-									new DefaultHttpClient().execute(new HttpPost(url));
-									locationDao.deleteDiveLocationLog(locationLog);
+									HttpResponse response = new DefaultHttpClient().execute(getHttpPost(locationLog));
+									if (response.getStatusLine().getStatusCode() == 200) {
+										locationDao.deleteDiveLocationLog(locationLog);
+									}
 								} catch (Exception e) {
-									Log.d(TAG, "Could not connect to " + url, e);
+									Log.d(TAG, "Could not connect to server for log " + locationLog, e);
 									locationDao.addDiveLocationLog(locationLog);
 									runOnUiThread(new Runnable() {
 										public void run() {
@@ -215,27 +213,17 @@ public class GpsActivity extends Activity {
 			new Thread(new Runnable() {
 				public void run() {
 					int success = 0;
-					boolean urlFailure = false;
 					// Send locations
 					for (int i = 0; i < locations.size() && !cancel.get(); ++i) {
 						DiveLocationLog log = locations.get(i);
-						String url = getSendUrl(log);
-						if (url == null) { // Could not build URL, show error
-							runOnUiThread(new Runnable() {
-								public void run() {
-									Toast.makeText(GpsActivity.this, R.string.error_no_settings, Toast.LENGTH_SHORT).show();
-								}
-							});
-							urlFailure = true;
-							break;
-						} else {
-							try {
-								new DefaultHttpClient().execute(new HttpPost(url));
+						try {
+							HttpResponse response = new DefaultHttpClient().execute(getHttpPost(log));
+							if (response.getStatusLine().getStatusCode() == 200) {
 								locationDao.deleteDiveLocationLog(log);
 								++success;
-							} catch (Exception e) {
-								Log.d(TAG, "Could not connect to " + url, e);
 							}
+						} catch (Exception e) {
+							Log.d(TAG, "Could not connect for log " + log, e);
 						}
 						// Update progress
 						Message msg = handler.obtainMessage();
@@ -248,15 +236,13 @@ public class GpsActivity extends Activity {
 					msg.arg1 = locations.size();
 					handler.sendMessage(msg);
 
-					if (!urlFailure) {
-						final int successCount = success;
-						final int totalCount = locations.size();
-						runOnUiThread(new Runnable() {
-							public void run() {
-								Toast.makeText(GpsActivity.this, getString(R.string.confirmation_location_sent, successCount, totalCount), Toast.LENGTH_SHORT).show();
-							}
-						});
-					}
+					final int successCount = success;
+					final int totalCount = locations.size();
+					runOnUiThread(new Runnable() {
+						public void run() {
+							Toast.makeText(GpsActivity.this, getString(R.string.confirmation_location_sent, successCount, totalCount), Toast.LENGTH_SHORT).show();
+						}
+					});
 				}
 			}).start();
     	}
